requirements.txt
discord.py
ffmpeg - remember to add it to your PATH
PyNaCl
dotenv
python-dotenv
youtube-dl 
```